+++ source/tensors/model/class_ising_tf_rf.cpp
@@ -49,7 +49,7 @@ void class_ising_tf_rf::set_parameters(TableMap &parameters) {
     h5tb.param.h_stdv   = std::any_cast<double>(parameters["h_stdv"]);
     h5tb.param.h_rand   = std::any_cast<double>(parameters["h_rand"]);
     h5tb.param.h_pert   = std::any_cast<double>(parameters["h_pert"]);
-    h5tb.param.spin_dim = std::any_cast<size_t>(parameters["spin_dim"]);
+    h5tb.param.spin_dim = std::any_cast<long>(parameters["spin_dim"]);
     std::strcpy(h5tb.param.distribution, std::any_cast<std::string>(parameters["distribution"]).c_str());
     if(h5tb.param.J2 != 0.0) throw std::runtime_error("Use of [J2] - Next-nearest neighbor coupling - is not implemented yet");
     all_mpo_parameters_have_been_set = true;
@@ -172,7 +172,7 @@ Eigen::MatrixXcd class_ising_tf_rf::single_site_hamiltonian(size_t position, siz
 
 std::unique_ptr<class_mpo_site> class_ising_tf_rf::clone() const { return std::make_unique<class_ising_tf_rf>(*this); }
 
-size_t class_ising_tf_rf::get_spin_dimension() const { return h5tb.param.spin_dim; }
+long class_ising_tf_rf::get_spin_dimension() const { return h5tb.param.spin_dim; }
 
 Eigen::Tensor<Scalar, 1> class_ising_tf_rf::get_MPO_edge_left() const {
     Eigen::Tensor<Scalar, 1> ledge(3);
